requirements.txt
fastapi
uvicorn
pydantic
streamlit
altair
google-generativeai
python-dotenv
langchain-core
jinja2
scikit-learn
pandas
numpy
scipy
loguru
pypdf
python-docx
aiofiles
python-magic
requests
pyyaml
black
pytest
pytest-asyncio
httpx
python-multipart
